Created new GridEntry control for setting the parameters for Image Grids. Currently used on Decks and Tokens.
