+++ Scripts/ImageManagement/IconLibrary.cs
@@ -147,6 +147,17 @@ public class IconLibrary : Dictionary<string, IconEntry>
             button.AddItem(icon, id);
             id++;
         }
+        
+        //put in user-image block
+        if (ProjectService.Instance.CurrentProject.Images.Count > 0)
+        {
+            button.AddSeparator();
+            foreach(var img in ProjectService.Instance.CurrentProject.Images)
+            {
+                button.AddItem($"u:{img.Value.Name}", id);
+                id++;
+            }
+        }
 
         button.AddSeparator();
